custom.lds: sort .text .tbl data with SORT_BY_INIT_PRIORITY()
SORT() sorts lexigraphically, we want to ensure to
sort by a priority number which is exactly what
SORT_BY_INIT_PRIORITY() seems to be used for that
purpose, so use that instead -- note this still requires
some vetting from source code on binutils. Its also
still not clear if since when SORT_BY_INIT_PRIORITY() is
available though. If SORT() is older, what does that
buy us in terms of binutils requirements ? SORT() is
already used in the Linux kernel [0], so if we a solution
that works there perhaps work off of just SORT().

For now we'll work with SORT_BY_INIT_PRIORITY() though
to at least annotate we might want a change. Once homework
is done to clearly distinguish the gains over SORT() we
can decide.

[0] See arch/x86/kernel/vmlinux.lds

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>