Add dedicated .tbl section
The default linker script will create a section for each
declared .tbl init routine we add, we can't sort these
this way. We also can't see the 00 and 99 tbls, the linker
probably optimizes this out as its all 0'd out anyway but
we need to be sure about the values of those. By adding a
dedicated section .tbl for all .tbl* stuff we can dump it
out with objdump -x -j .tbl main

...
                  CONTENTS, ALLOC, LOAD, DATA
SYMBOL TABLE:
0000000000601050 l    d  .tbl   0000000000000000              .tbl
0000000000601060 l     O .tbl   0000000000000000		__table_entries.2938
0000000000601060 l     O .tbl   0000000000000000		__table_entries.2940
0000000000601060 l     O .tbl   0000000000000000		__table_entries.2942
0000000000601060 l     O .tbl   0000000000000000		__table_entries.2944
0000000000601058 g     O .tbl   0000000000000008		bar_init_fn
0000000000601060 g     O .tbl   0000000000000008              x_init_fn
0000000000601050 g     O .tbl   0000000000000008		foo_init_fn
0000000000601070 g       .tbl   0000000000000000		__bss_start
0000000000601068 g     O .tbl   0000000000000008              y_init_fn
0000000000601050 g     O .tbl   0000000000000000              .hidden __TMC_END__
...

Its not so obvious that the the __table_entries.29* are the
00 and 99 .tbl entry references though.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>