custom.lds: move .tbl back out to its own section
Now that we've demoed the binary let's stick to something
that would look more like we would use upstream on Linux.

We'd want its own section, perhaps right after .iommu_table
or eventually extending / enhancing it by making it more
generic.

For now we keep SORT_BY_INIT_PRIORITY() but annotate that
we may just want SORT() for now. While at it, add an ALIGN(8).

Now the fun begins -- extending our init struct to match
more of our needs. This will be to match requirements of the
kernels' iommu_table hacks (dependencies), but we're also
shooting to make it very generic.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>