init: and add critical annotation support
Add a critical boolean entry for each init struct. This lets
init sequences annotate whether or not they are required to
succeed completion before letting init move on with life.

This deviates from the kernel's iommu init sequence stuff, it
uses void for the init sequence routines. We have to modify our
init routine callback to return an int, while we're doing that
just rename the callback to early_init as our next change will
be to add support for a late init.

For now we note that X is the only thing pegged as being critical.
If we add this to the kernel the iommu stuff won't need critical
bools, if they don't have it we'll just continue on as before. This
will be more useful for other areas of the kernel.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>