fix annotation a bit
Well it seems the sort dependency thing actually works.
I can see this if I move the driver to INIT_EARLY and
keep the pci driver on INIT_NORMAL.

If this is done then the driver would run first. This is
obviously wrong, so to fix that we could add an annotation
of having the driver depend on detect_pci, we leave that
commented out first to let you tests and see that for
yourself.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>