make the driver depend on pci
What does not work is if pci fails we should not move on with
PCI drivers, we can try to fix that next. What this should mean is
that if an init sequence has a "depends" and if its marked as
critical then it cannot allow things that depend on it to
continue. Right now if an init sequence is listed as critical
we'll fail the entire init sequence right then where it failed,
we want to enable a middle ground, in such a way to be express
that one should not let things that depend move forward unless
it continued successfully.

Note that in this case the driver had a higher init priority
though originally:

driver: INIT_EARLY
pci: INIT_NORMAL

 #define INIT_EARLY      01
 #define INIT_SERIAL     02
 #define INIT_CONSOLE    03
 #define INIT_NORMAL     04

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>