add pci
Forgot to add this to the tree. Note that if
we make PCI critical, since its failing right
now we end up with:

Completed initializing Foo thing !
Initializing PCI buses ...
Early init failed

So all of init fails. If we remove the critical
annotation we get:

Initializing PCI buses ...
Failed to initialize PCI buses on early init, but its not critical
Initializing ACME(TM) Driver ...
Completed initializing ACME(TM) Driver !

We want to add support for ACME driver to be able to
annotate it depends on PCI subsystem for this simple
example case but to let the init sequence go on.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>