xen: use startup_xen() for init
Mimic Linux code for xen init. This is the entry point used
from handoff by the Xen Hypervisor. The way the hypervisor
finds finds the entry point is by inspecting the Linux binary,
lookng for the XEN_ELFNOTE_ENTRY elf note and using that.

For bare metal and lguest the entry points are different, and
we'll mimic that next.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>