init: start using hardware_subarch as an example
This starts to make use of the hardware_subarch from the
boot_params, available as of the 2.07 x86 boot protocol
to illustrate how it can be used on early init.

X86_FEATURE_HYPERVISOR cannot be used before setup_arch()
as its only available after early_cpu_init(), we want
something that will available as early as the kernel
boots. X86_FEATURE_HYPERVISOR is also very generic.

The hardware_subarch has been used for other x86 subarchs
and at least for 32-bit there is already even a switch for
it on i386_start_kernel().

We keep this code in place for now as we are evaluating
the use complexities of settings this value for the different
boot loaders: xen / qemu / grub2, etc. We also want something
that will be actually be useful for not only xen but also kvm,
so further evaluation of possible dead code for kvm is also
beind done.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>