parse-bzimage: add bzimage parser
Prior to booting Linux qemu / xen / grub2 needs to parse
the bzimage, they also then re-use some of the same to set
the 0 page (on 32-bit, on 64-bit this can be anywhere) for
Xen HVM and KVM.

This code is added to aid evaluaton of making changes on
qemu / xen / grub / native kvm tool, etc. I at least have
patches ready now for:

  * qemu - supports kvm and Xen HVM
  * native kvm tool

I'm still evaluating Xen's requirements for PV and PVH, this
is a bit complex due to the requirements of having a custom
boot loader use when one doesn't want to specify the kernel
and initrd manually. For that see:

http://wiki.xenproject.org/wiki/PvGrub2
http://xenbits.xen.org/docs/unstable/misc/x86-xenpv-bootloader.html

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>