init: enable support for no depends() callback
The Linux kernel's IOMMU init table framework requires
a depends() callback to check if your feature is detected
or needed. Since the GPXE init solution works with a linker
script to help you sort the init routines by levels we don't
need to require a depends callback for all init routines,
if they don't have one we skip re-sorting them based on
this heuristic.

This will be more useful as we also add support later for
subarchitecture dependency semantics, which we'll do next.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>