custom.lds.S, tables.h: make tables subsystem specific
The custom.lds.S was still working with a generic table,
update that to make it clear how one can work with different
subsystems. This deviates from gPXE's own solution slightly
in that each table needs to be declared separately. That
works better for Linux given we're combining the IOMMU init
solution as well which has its own sorting options with
memmove(). It also works better for Linux given the size
of Linux and how an issue on one subsystem could affect
others. We want to compartamentalize each table solution
separateley.

While at it, revamp tables.h with our own Linux documentation.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>